// src/store/types.ts
export enum TYPES {
  FETCH_CURRENCY_LIST = "FETCH_CURRENCY_LIST",
  SET_VALUE_NUMBER = "SET_VALUE_NUMBER",
  SET_CONVERTED_VALUE = "SET_CONVERTED_VALUE",
  SET_CURRENCY_FROM = "SET_CURRENCY_FROM",
  SET_CURRENCY_TO = "SET_CURRENCY_TO",
  RELOAD_STATE = "RELOAD_STATE",

  SET_LOADING = "SET_LOADING",
  SET_ERROR = "SET_ERROR",
}